登录显示信息

+++ src/main/java/com/nowcoder/community/dao/UserMapper.java
@@ -7,12 +7,12 @@ import org.apache.ibatis.annotations.Param;
 
 @Mapper
 public interface UserMapper {
-    User selectById(@Param("id") Long id);
+    User selectById(@Param("id") Integer id);
     User selectByUsername(@Param("username") String username);
     User selectByEmail(@Param("email") String email);
     int insertUser(User user);
-    int updateStatus(@Param("id") Long id, @Param("status") Integer status);
-    int updateHeaderUrl(@Param("id") Long id, @Param("headerUrl") String headerUrl);
-    int updatePassword(@Param("id") Long id, @Param("password") String password);
+    int updateStatus(@Param("id") Integer id, @Param("status") Integer status);
+    int updateHeaderUrl(@Param("id") Integer id, @Param("headerUrl") String headerUrl);
+    int updatePassword(@Param("id") Integer id, @Param("password") String password);
 }
